Parameters only saved if events exist, if csv file exists it is not recreated and refactor to parameter select page

+++ src/main/java/com.smithies/analoguedataexporter/controllers/EventsController.java
@@ -18,7 +18,7 @@ public class EventsController {
 
     @GetMapping(path="/{id}")
     public @ResponseBody
-    AnalogueEvent getEvent(@PathVariable("id") final Integer id) {
+    AnalogueEvent getEvent(@PathVariable("id") final long id) {
         AnalogueEvent analogueEvent = analogueEventRepository.findOne(id);
         return analogueEvent;
     }
@@ -26,7 +26,7 @@ public class EventsController {
     @GetMapping(path="")
     public @ResponseBody
     Iterable<AnalogueEvent> getEvents() {
-        List<Integer> eventIds = Arrays.asList(685144355, 685144356, 685144353);
+        List<Long> eventIds = Arrays.asList(685144355l, 685144356l, 685144353l);
         return analogueEventRepository.findAll(eventIds);
     }
